+++ demo/demo/src/main/java/com/example/demo/repository/ReceitaRepository.java
@@ -1,10 +1,10 @@
 package com.example.demo.repository;
 
-import com.example.demo.entity.ReceitaEntity;
+import com.example.demo.model.entity.ReceitaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReceitaRepository extends JpaRepository<ReceitaEntity, Long> {
+public interface ReceitaRepository extends JpaRepository<ReceitaEntity, Integer> {
 
 }
